+++ app-backend/database/src/main/scala/com/azavea/rf/database/fields/ImageFields.scala
@@ -3,7 +3,7 @@ package com.azavea.rf.database.fields
 import com.azavea.rf.database.ExtendedPostgresDriver.api._
 
 trait ImageFields  { self: Table[_] =>
-  def rawDataBytes: Rep[Int]
+  def rawDataBytes: Rep[Long]
   def filename: Rep[String]
   def sourceuri: Rep[String]
 }
